praefect: Praefect should not hang if db is not reachable
If the database is not accessible the at the time of Praefect
start it hangs forever and doesn't report much information
about the reason. The attempt to fix it by using PingContext
with timeout set on the context failed because of the issue
in the lib/pq that is not yet fixed. The problem is solved
by a separate goroutine that issues the PingContext and main
goroutine watching the timeout. If timeout occurs it returns
an error up to the caller. Another goroutine won't be over
and hangs, but we don't care much as it is a critical error
and the service will be terminated.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3678