helper: Drain timer ticker on reset
Resetting `time.Timer`s is quite tricky in cases where the tick hasn't
been consumed by the caller given that one first has to stop the ticker,
then consume the event in case there is any and finally the ticker can
be reenabled by calling reset. Given that this is not quite obvious
without taking a closer look at `time.Timer`'s documentation this is
easy enough to miss.

Adjust our own `Reset()` wrapper in the helper package to do this dance
such that `Reset()` will reset the timer in all cases.