repository: Convert `ReplicateRepository()` to atomically create repos
When the target repository for `ReplicateRepository()` doesn't yet
exist, then we will create it by first initializing an empty repository
and then fetching and extracting an archive of the source repository.
This currently happens in a non-atomic way, so if any other RPC call
concurrently modifies the target repository then the end state is
indeterminate.

Now that we have a central helper function which knows to create
repositories atomically, convert `ReplicateRepository()` to use this
new helper. This change makes the case where no repository exists yet
completely atomic: we either end up with the target repository in place
if the RPC call finished, or alternatively it will not have been created
at all.

Changelog: fixed