git2go: Use gob for conflicts sub-command invocation
So far we used JSON serialization to invoke conflicts
sub-command of the git2go executable. We are moving
towards usage of the gob serialization to do the same
and use stdin and stdout of the process to get an input
and return the result of the execution. gob doesn't
require us to do mapping of the fields and allows
passing typed errors between the processes.

Some functions were re-ordered to group them
by usage: methods of the same struct, exportable
functions, private functions.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3232