git2go: Fix error type check
The error check in one of the code blocks
was done incorrectly. This change fixes the
error conversion logic done improperly by using
a pointer of git.GitError type as a destination
as git2go lib returns a pointer type not a value.
The other fix is applied to the if condition check.
Now it breaks in case the code matches expected and
returns an error if not.