catfile: Move remaining bytes into object
The catfile object reader needs to track the number of bytes remaining
to be read such that we know when it's safe to start reading another
object, and when the catfile reader is dirty and thus cannot be returned
to the cache. These bytes are currently stored in the reader itself, and
as a result the object reader has to reach into the parent reader to
adjust bytes in there. This violates the separation of concrens, and
furthermore it results in locking issues when one wants to refactor the
reader.

Fix the design issue by instead moving the remaning byte count into the
object itself. The parent reader instead keeps track of the current
object and asks it for how many bytes are left to be read.

No change in behaviour is expected from this change, but it only serves
to disentangle locking for the upcoming object reader queue.