conflicts: Backwards compatibility
With introduction of the gob command parameters
passing we should make sure the command is backwards
compatible with old approach where the flag is used
to pass parameters. It is needed as rollout is not
an atomic and we may be in situation when the old
version of gitaly invokes new version of the conflicts
sub-command.