lstree: Split up parser and supporting data structures
The lstree package contains the somewhat weirdly named "last_commits.go"
file, which contains a parser for git-ls-tree(1) output. This stems from
the fact that the parser was originally used to implement the
`ListLastCommitsForTree()` RPC. The parser is much more generally useful
though, so it doesn't make sense to tie it to the RPC call.

Split up the code into two generically named files "parser.go" and
"entries.go" to help readers find what they search for.