catfile: Drop `finish()` function returned when creating traces
When creating traces, the caller must always make sure to finalize the
trace when it will get no more requests such that the corresponding span
is closed and that metrics are updated accordingly. We thus return a
callback from `startTrace()` which the caller is supposed to invoke at
some point in time. It's not immediately clear though whether the caller
must invoke the callback _and_ call the also-existing `finish()`
function, where both are in fact the same.

Refactor the code to not return the function from `startTrace()`
anymore to avoid this confusion and document that one must call
`finish()` manually. While at it, fix the documentation to not mention
the two separate contexts anymore: they have been removed.