catfile: Split out object info reader queue
Similar to the preceding commit, this commit splits out a new object
info reader queue such that we can eventually change callers to use
batched requests for object info.

No change in behaviour is expected given that the queue is not yet used
by anything except as implementation detail of the `Info()` function.