git: Unify fsck configuration
The fsck configuration needs to be injected into multiple commands,
where the prefix of the configuration key is different depending on
whether it's used by git-receive-pack(1) or git-fetch-pack(1), which is
why we have duplicated the configuration in two places. It's not easily
maintainable though and only a question of time until we forget to
update one of both.

Refactor the code to instead generate the configuration via a separate
function to deduplicate them.