testhelper: Add ability to run benchmarks on feature sets
Our `FeatureSets` test helper is only able to run normal tests: running
benchmarks is not something it can do given that we have a `testing.B`
available in that context, not a `testing.T`.

Add a new `Bench()` function which allows callers to run benchmarks with
feature flags. Ideally, we'd just adapt `Run()` to accept a `testing.TB`
instead, but `testing.TB` does not provide a `Run()` function and thus
cannot be used here.