lstree: Optimize allocations when parsing trees
Convert the lstree parser to use `ReadBytes()` per each item we want to
read instead of using it per-line and then using `bytes.SplitN()` to
split up the line. This is both easier to read and more efficient with
regards to memory allocations. Before this change:

    BenchmarkGetTreeEntries/all_features_enabled/recursive_from_root-16             7   158146209 ns/op   105882963 B/op   1202079 allocs/op
    BenchmarkGetTreeEntries/all_features_enabled/recursive_from_subdirectory-16   100    13859108 ns/op     4456390 B/op     53169 allocs/op

And after this change:

    BenchmarkGetTreeEntries/all_features_enabled/recursive_from_root-16             7   159637837 ns/op    97305347 B/op   1202069 allocs/op
    BenchmarkGetTreeEntries/all_features_enabled/recursive_from_subdirectory-16    86    12786680 ns/op     4083270 B/op     53172 allocs/op

This shows that we're using roughly 10% less memory per invocation of
`GetTreeEntries()`.

Changelog: performance