praefect: Create records for additional repositories in tests
Praefect's Coordinator contains a hack to create records for repositories
that are being accessed. This is done to avoid repository not found errors
when Gitaly tests are run with Praefect in front of the Gitaly. The tests do
not create the repositories via the API but instead place the repository directly
in the storage. Ideally, the tests would call CreateRepository to set up the test
state as needed so the hack would not be needed. As we are going to be rewriting
relative paths of additional repositories in the requests, the hack needs to be
expanded to create records for the additional repositories as well. Otherwise the
tests in Gitaly that use additional repositories fail.