praefect: allow check subcommand to be skipped
Since the startup checks can potentiall take a while, we want to allow
the option to skip the startup checks via an environment variable. This
way, the startup tooling such as omnibus can set this environment variable
when it needs to skip the startup healthchecks before starting Praefect.

Changelog: added