grpcstats: PayloadBytes to track payload bytes
PayloadBytes implements stats.Handler interfaces and
allows collect information about transmitted bytes
during gRPC method invocation. It covers both unary
and stream invocations. The HandleRPC method receives
specific type of stats depending on the stage of the
method invocation. The amount of bytes in received
payload extracted from the *stats.InPayload type.
The amount of bytes in sent payload extracted from
the *stats.OutPayload type. All statistics is gathered
into the PayloadBytesStats type instance that is stored
in the context. The initialization of the context with
instance happens on TagRPC method invocation.
The Fields method of the PayloadBytesStats returns logging
information that should be added to the logging context
for each request.
THe FieldsProducer used to extract info from the context
and will be used in upcoming commits.