log: Extended MessageProducer
To make logging more flexible this change adds
a new MessageProducer. It allows use passed in
MessageProducer as an actual logs writer and a
bunch of FieldsProducer(s). Each FieldsProducer
can add additional info into the logging context
before the log will be written. This is a preparation
change for adding more data into the logs.