log: PropagationMessageProducer to catch logging
The new PropagationMessageProducer creates a special
MessageProducer that doesn't produce any logs, but
instead populates a special messageProducerHolder
that is stored inside the context. This allows to
popup the logging up to the place where the interceptors
are not running. The context modification is a bad
practice, but we don't have much other options here.