catfile: Implement support for flushing
Our git-cat-file(1) processes currently use unbuffered I/O both to
accept requests, where we directly write into the process's pipe, and to
write out results given that we never pass the "--buffer" flag to the
processes. As a result, we can see in production systems that there's
quite some overhead at times of high load due to writing to these pipes,
which causes an excessive number of read/write syscalls and thus a lot
of context switches.

We couldn't really implement buffered I/O for this subsystem in the past
given that all requesting and reading objects was directly tied to each
other. This has changed now with the recent introduction of the request
queue, which allows us to split up requesting objects and reading the
results.

Implement support for flushing in the request queue as a first step
towards using buffered I/O. As of now this is still a no-op, but will be
extended in subsequent commits as we convert the code to use buffering.