catfile: Flush requests as required
While the request queue of catfile processes already exposes a way to
flush outstanding requests, calling it is not mandatory yet given that
we do not yet use buffered I/O. This is going to change in subsequent
commits though, so let's prepare for that by calling `Flush()` in all
required places.

Note that in case of the gitpipe package we're also converting the
channels to be buffered channels: this allow us to queue multiple
requests before flushing such that we can make better use of the
upcoming buffered stdin.