catfile: Use buffered mode in git-cat-file(1)
The git-cat-file(1) command can be asked to optionally buffer all writes
to stdout to reduce the number of write syscalls it needs to execute.
This can be much more efficient, especially when requesting a large
number of objects. We haven't been able to use that mode in the catfile
package though given that it has been request-response based in the
past, where we always tried to consume the object immediately after it
has been requested. This limitation has been lifted with the
introduction of the request queue, which allows callers to queue
requests and read the results in different Goroutines. Furthermore, with
the introduction of flushing semantics, we can now also ask Git to flush
any not-yet-written bytes to stdout.

Unfortunately, asking git-cat-file(1) to flush isn't as straight-foward
as one would wish though: it doesn't have any way to ask it to flush.
Luckily, we can use a workaround though: when git-cat-file(1) is asked
for an object which doesn't exist, then it writes an error message
"$object missing" to its stdout and then immediately flushes stdout.
This can thus be used as an easy primitive to implement flushing, as
long as one filters out the error message on the other side such that it
doesn't ever leak to users of the package.

The only remaining question is what object name to use for flushing. It
must be an object which is knnown to never exist, so any valid reference
name or object ID wouldn't work in this context. The easiest object that
cannot exist would be the empty string, but if this ever leaks to the
caller somehow then it would be rather confusing. Instead, we take
advantage of the fact that refs cannot ever contain whitespace and use
"\tFLUSH\t" as the command: it shouldn't ever point to a valid object,
and furthermore it also gives a hint why it actually exists to any
unsuspecting user who happened to take a glimpse of this string.

Convert the code to use buffered mode in git-cat-file(1) and amend our
`Flush()` method to make use of our hack.

Changelog: performance