log: PerRPCLogHandler for extended logging
As we are required to collect information about
payload size of each RPC this is not something
that could be done with interceptors, but we
can benefit from the stats.Handler that provides
access to this info. The problem is that the Handler
wraps all interceptors, so the logger that is set
by the interceptor on the context is not accessible
to it. To get around that we should use PerRPCLogHandler
in combination with PropagationMessageProducer.
The latter catches the log parameters and the logger
itself and propagates it into the context. Once the
RPC is over the PerRPCLogHandler will use caught log
parameters and logger to produce actual log with any
additional parameters that were collected by FieldProducers.