grpcstats: Extend log with payload size
This change finalize introduction of the payload
size into the logging. The PerRPCLogHandler now
responsible for writing log messages produced by
the logrus interceptor with the additional data
gathered by the PayloadBytes (payload sizes).
The (Unary|Stream)LogDataCatcherServerInterceptor
catches log data produced by the logrus interceptor
and set it on the holder introduced in the previous
commits. So it will be available for the PerRPCLogHandler
once all interceptors are done.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3867

Changelog: added