catfile: Use buffered writes to queue requests
When requesting objects or object info from the request queues, we're
writing the request directly into the command's pipe. This thus amounts
to one syscall per request, which is excessive and causes problems in
production due to the resulting context switches.

Convert the code to instead use a buffered writer to batch several
requests into a single write. Flushing the request queue is extended to
also flush out this buffered writer.

Changelog: performance