remove-repository: A new sub-command for the praefect to remove repository
Users of the praefect are missing tools to manage state of the cluster.
One of such tools is a repository removal cli. It should be used to
remove any repository from the cluster. The removal covers cleanup of
the database and removal from the gitaly storages. The command tries
to remove as much as possible first by removing from praefect, replication
event queue and each gitaly node configured in the provided config file.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3771

Changelog: added