list-untracked-repositories: Praefect sub-command to show untracked repositories
The change is a backport of the functionality implemented
to resolve #3792
issue. A new sub-command for the praefect binary. On run it
connects to all gitaly storages set in the configuration file
and receives from each list of the repositories existing on
the disk. Each repository then checked if it exists in the
praefect database and if it is not the location of the
repository is printed out in JSON format to the stdout of the
process.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3792

Changelog: added