Move WaitForQueries helper to glsql
WaitForQueries is a helper function that is used when testing concurrent
query execution. It allows for checking in the DB there is a concurrent
query executing rather than relying on sleep. This will be needed to test
the concurrent updates in HealthManager. This commit moves the helper to
glsql where it can be imported by any test that needs it.