ruby: Remove unused code
Due to our piecemeal conversions from Ruby to Go, it's hard at times to
find Ruby parts which are not used anymore. As a result, we have
accumulated quite a lot of baggage which was missed to be removed. While
this is not much of a problem by itself, it does create tension with the
upcoming change to support deployments with bundled Git versions, where
we don't have a complete Git installation anymore but only a small set
of binaries.

Remove unused code to make the transition easier.