objectpool: Fix data race with logger
In our test to verify that we correctly collect objectpool statistics
via the logger, we create our own logger with a backing buffer. This
buffer is being accessed concurrently though, once by the gRPC
middleware which tries to write into it and once from the test itself
after the RPC call has finished to verify it contains what we expect it
to contain. Given that we don't use locking for this the result is a
data race.

Fix the data race by instead using a NullLogger and extracting the log
entries via a hook. This also allows us to tighten our checks to assert
we see all expected stats.