Fix concurrent execution tests of PerRepositoryElector
PerRepositoryElector's tests contain code to ensure concurrent
execution of two transactions for each test. However, the second
transaction actually blocks on inserting the health records, not on
performing the election query. As such, the second election query
can see the changes done by the first query after it is committed.
This commit fixes the issue in the test suite by setting the healthy
nodes ahead of creatin the transactions and ensuring the second transaction
is actually blocked on the election query. This way it's guaranteed
the second election query doesn't yet see the changes performed by the
first election query. The tests fail as the second query actually performs a
second failover which is not intented.