Add config value MetatdataMonitoringEnabled
With multiple Praefect instances running, each one will hit the database
for metrics. This causes unnecessary load on the database and in some
cases, fatally so. To fix this, we will add a new config value so that
we can control whicih Praefect serves as thei one that will monitor the
database.

Changelog: added