localrepo: Add GuessHead to guess what HEAD would be given a reference
Since bundles do not store symbolic refs we need to guess what the
branch would be given the reference in the bundle. Ideally we would rely
on git for this, but git will only set the default branch on git-clone
which we cannot use when apply multiple incremental backup bundles. We
can however simulate the same function.