ruby: Silence spec warnings when there is no template dir
Our rspecs raise a bunch of warnings when creating repositories about
missing template directories if the template directory is indeed missing
from the host it's running on. We do not care about templates though and
have moved away from them already, where Go already knows to not use
them at all anymore. The upcoming change to bundle Git with Gitaly will
make it the norm that there are no templates anymore, and thus we'd
always start to hit this warning.

Fix the issue by explicitly telling Git that it shouldn't use the
template dir when initializing test repos.