testhelper: Drop Ruby flags from default feature set constructor
Go-specific and Ruby-specific feature flags have different prefixes,
which is why the feature set constructor accepts two lists of flags for
both areas. Ruby feature flags are not currently in use by anything
though, and they likely will not be for the foreseeable future. But they
do make the calling convention for the feature set constructor
needlessly verbose.

Refactor the helper to only accept Go feature flags. A new function
`NewFeatureFlagsWithRubyFlags()` is created in case we'll ever have to
use Ruby feature flags.