commit: Do not raise error when listing tree entries for nonexistent ref
The GetTreeEntries RPC has weird calling semantics: when the caller
requests entries for a nonexisteng revision or path, then it won't error
but instead just returns an empty set. While the rewrite for recursive
listings introduced via 08507b227 (commit: Convert GetTreeEntries to use
git-ls-tree(1), 2021-11-08) correctly honors the case when trying to
list entries for an object which is not a treeish, it fails to do the
same when using an invalid revision. As a result, the new code returns
errors where we previously just silently failed.

Fix this by swallowing `lstree.ErrNotExist` errors.

Changelog: fixed