Don't demote primaries
Praefect currently demotes a primary if the consensus is that it isn't
healthy. There's no real reason to do this though. It causes an extra
record update and it doesn't really improve the situation in anyway.
This commit removes the demotion behavior and keeps the primary in place
if there are no valid candidates for promotion. This allows for Praefects
to keep accepting write requests for the primary if they have a healthy
connection to it locally.

Changelog: changed