Remove HealthConsensus querying from HealthManager
HealthManager previously used to query for the health consensus and
see if it has changed. If it had, it signaled on a channel that would
then trigger an election run. We've since moved on to lazy elections
which do not require this sort of eventing. This commit removes the
consensus querying as it is not used for anything anymore.

The tests of the health consensus are still left in place though. The
healthy_storages view and it's logic still needs to be tested and this
is the most natural place to do so.