Simplify HealthManager first update send
HealthManager is now only sending an update on the first update. This
notification is used by Praefect in the main to ensure the first round
of health checks have been performed before proceeding to start the
servers. As we only ever send to the channel once and it is buffered,
this simplifies the send by removing the default case as this never
blocks.