Rewrite relative paths in backup restoration test
The backup restoration tests reach into the filesystem to setup test
state and to verify the backup was correctly restored. These direct
accesses to the repository via the filesystem will break when Praefect
begins generating unique relative paths for repositories as the repository
is no longer at relative path where the tests expect it to be. To keep the
tests functioning, this commit resolves the replica path of the repository
if the test is ran with Praefect in front of the test.