Improve custom hook error logging
Previously if a custom hook failed, it would be difficult to determine
which hook in question were actually failing without using strace. The
failing hook could be global or a project hook. We now log the filename
along with the error message.

Relates to https://gitlab.com/gitlab-org/gitaly/-/issues/2923

Changelog: added